+++ CHANGELOG.md
@@ -4,7 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project generally adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.11]
+
+### Added
+
+- NEB runner
+- NEB-based TS job for NewtonNet
+
 ## [0.11.10]
+
+### Fixed
+
+- Fixed compatability with Prefect 3.1.11
 
 ## [0.11.9]